(DOCSP-57780) Adds the v1.52.1 release (#18189)
https://jira.mongodb.org/browse/DOCSP-57780


[STAGE](https://deploy-preview-18189--mongodb-atlas-cli.netlify.app/docs/atlas/cli/upcoming/atlas-cli-changelog/#atlas-cli-1.52.1)

+++ content/tools/atlas-cli-commands-toc/generate-cli-commands.ts
@@ -274,7 +274,7 @@ async function copyCommandFilesFromGit(tagOrBranch: string): Promise<string[]> {
         const lowerFile = file.toLowerCase();
         // Do not delete kubernetes command files or deployment command files
         if (fs.statSync(filePath).isFile() && !validFiles.has(file) && 
-            !lowerFile.includes('kubernetes') && !lowerFile.includes('deployment')) {
+            !lowerFile.includes('kubernetes') && !lowerFile.includes('deployment') && !lowerFile.includes('local')) {
           fs.unlinkSync(filePath);
           console.log(`🗑️ Deleted extra file: ${filePath}`);
         }
